requirements.txt
Django~=1.11.5
django-environ~=0.4.4
django-extensions~=1.9.1
django-podcast-feed~=0.0.2
huey~=1.4.1
iso8601~=0.1.12
psycopg2~=2.7.3
pytz==2017.2
redis~=2.10.6
requests~=2.18.4
waitress~=1.0.2
whitenoise~=3.3.0
